CLI-837 Fold migration.ts into claude-hooks-migration.ts
