+++ CLAUDE.md
@@ -50,7 +50,7 @@ The bare `sonar analyze` command accepts `-p, --project <project>` for the agent
 
 The declarative-integration engine lives under `src/core/framework/` — it has no CLI-specific coupling, so it sits alongside `src/core/host/` rather than nested inside it: `src/core/framework/features/` (types, selection, installer, install-integration, feature-target, install-preview, completion-summary, installation-recorder, registry, reconcile), `src/core/framework/resources/`, and `src/core/framework/dependencies/` (including the generic SonarSource binary dependency) as sibling modules. New integration descriptors should import from `@/core/framework/features` for dependency/resource factories, operations, and registry validation. Command handlers should keep command-specific validation, prompts, and target resolution thin, then delegate feature selection, generic install messages, dependency/resource application, and state recording to `src/core/framework/features/install-integration.ts`. Only the CAG binary dependency declaration stays in commands, at `src/commands/integrate/_common/context-augmentation-dependency.ts`, since it calls `SonarQubeClient` and CLI output helpers.
 
-The self-update lifecycle lives under `src/core/update/` — a top-level peer of `core/host/` and `core/framework/`, not nested inside either, because it legitimately depends on both: `check.ts`/`notification.ts` (version check and the post-command update notice), `post-update.ts` (the orchestrator run once per upgrade from `src/index.ts`), `migration.ts` (obsolete Claude Code hook artifact cleanup — also called directly by `commands/integrate/claude/` when a user manually re-runs the integration, not just from the post-update flow, since both are the same underlying migration concern), and three single-purpose migration modules `post-update.ts` calls in sequence — `telemetry-migration.ts`, `claude-hooks-migration.ts`, `binary-refresh.ts`. `post-update.ts` also defines `migrateDeclarativeIntegrations` directly (not worth its own file — it's a 4-line load/save wrapper) around `reconcileInstalledIntegrations` (`src/core/framework/features/reconcile.ts`), which owns the actual replay-declared-features-against-recorded-state logic — kept in framework rather than host/update so nothing under `core/host/` ever needs to import from `core/framework/`.
+The self-update lifecycle lives under `src/core/update/` — a top-level peer of `core/host/` and `core/framework/`, not nested inside either, because it legitimately depends on both: `check.ts`/`notification.ts` (version check and the post-command update notice), `post-update.ts` (the orchestrator run once per upgrade from `src/index.ts`), and three single-purpose migration modules `post-update.ts` calls in sequence — `telemetry-migration.ts`, `claude-hooks-migration.ts`, `binary-refresh.ts`. `claude-hooks-migration.ts` also owns the sonar-a3s obsolete-hook-artifact cleanup helpers (`OBSOLETE_A3S_MARKER`, `removeObsoleteHookArtifacts`, `cleanObsoleteFromState`, `migrateHookScripts`) — they're Claude-specific (hardcoded to `.claude`, not parameterized over agent or marker) and called directly by `commands/integrate/claude/index.ts` (manual re-run) and `post-update.ts` (top-level state cleanup) in addition to `migrateClaudeCodeHooks` in the same file, so a separate `migration.ts` wasn't earning its keep as its own module. `post-update.ts` also defines `migrateDeclarativeIntegrations` directly (not worth its own file — it's a 4-line load/save wrapper) around `reconcileInstalledIntegrations` (`src/core/framework/features/reconcile.ts`), which owns the actual replay-declared-features-against-recorded-state logic — kept in framework rather than host/update so nothing under `core/host/` ever needs to import from `core/framework/`.
 
 Resource removers delete an emptied file instead of leaving a husk: `text-snippet` when only whitespace remains, patch removers (`json-patch`/`yaml-patch`/`toml-patch`) when the document prunes (via `pruneEmptyContainers`) to its `defaultValue`. User/other-feature content survives pruning and keeps the file; empty parent dirs are left in place.
