+++ src/commands/integrate/claude/index.ts
@@ -25,7 +25,7 @@ import { homedir } from 'node:os';
 import { installIntegration } from '@/core/framework/features';
 import type { ResolvedAuth } from '@/core/host/auth-resolver.ts';
 import type { IntegrationStateAttribute } from '@/core/state/state.ts';
-import { OBSOLETE_A3S_MARKER, removeObsoleteHookArtifacts } from '@/core/update/migration.ts';
+import { removeObsoleteHookArtifacts } from '@/core/update/claude-hooks-migration.ts';
 
 import { printAgentNonInteractiveAlternativeHint } from '../../_common/agent-prompt-hint.ts';
 import {
@@ -108,7 +108,7 @@ export async function integrateClaude(
   } catch (error) {
     installError = error instanceof Error ? error : new Error(String(error));
   }
-  await removeObsoleteHookArtifacts(ctx.project.rootDir, OBSOLETE_A3S_MARKER);
+  await removeObsoleteHookArtifacts(ctx.project.rootDir);
   if (installError) {
     throw installError;
   }
