+++ tests/unit/core/update/claude-hooks-migration.test.ts
@@ -19,22 +19,44 @@
  */
 
 import * as fs from 'node:fs';
+import { tmpdir } from 'node:os';
+import { join } from 'node:path';
 
 import { afterEach, beforeEach, describe, expect, it, Mock, spyOn } from 'bun:test';
 
 import { CLAUDE_INTEGRATION_ID } from '@/commands/integrate/claude/declaration.ts';
 import * as hooks from '@/commands/integrate/claude/hooks.ts';
-import type { CliState, HookExtension } from '@/core/state/state.ts';
+import type { AgentExtension, CliState, HookExtension } from '@/core/state/state.ts';
 import { getDefaultState } from '@/core/state/state.ts';
 import * as stateRepository from '@/core/state/state-repository.ts';
+import { setMockUi } from '@/core/ui';
+import * as migration from '@/core/update/claude-hooks-migration.ts';
 import { migrateClaudeCodeHooks } from '@/core/update/claude-hooks-migration.ts';
-import * as migration from '@/core/update/migration.ts';
 
 const FAKE_HOME = '/fake/home';
 const homedirFn = () => FAKE_HOME;
+const OLD_VERSION = '0.4.0';
+// Mirrors the private OBSOLETE_A3S_MARKER constant in claude-hooks-migration.ts.
+const OBSOLETE_A3S_MARKER = 'sonar-a3s';
 
 function makeState(): CliState {
   return getDefaultState('1.0.0');
+}
+
+function seedAgentExtension(state: CliState, extension: AgentExtension): void {
+  const idx = state.agentExtensions.findIndex(
+    (e) =>
+      e.agentId === extension.agentId &&
+      e.projectRoot === extension.projectRoot &&
+      e.kind === extension.kind &&
+      e.name === extension.name &&
+      (e.kind !== 'hook' || extension.kind !== 'hook' || e.hookType === extension.hookType),
+  );
+  if (idx >= 0) {
+    state.agentExtensions[idx] = { ...extension, id: state.agentExtensions[idx].id };
+  } else {
+    state.agentExtensions.push(extension);
+  }
 }
 
 function makeStateWithExtensions(extensions: HookExtension[], configured = true): CliState {
@@ -298,7 +320,7 @@ describe('migrateClaudeCodeHooks', () => {
     expect(actual).toBeUndefined();
   });
 
-  it('calls removeObsoleteHookArtifacts once per location with the sonar-a3s marker', async () => {
+  it('calls removeObsoleteHookArtifacts once per location', async () => {
     const state = makeStateWithExtensions([
       makeExtension('/proj/alpha', false),
       makeExtension('/proj/beta', false),
@@ -308,13 +330,250 @@ describe('migrateClaudeCodeHooks', () => {
     await migrateClaudeCodeHooks(hooks.installHooks, CLAUDE_INTEGRATION_ID, homedirFn);
 
     expect(removeObsoleteHookArtifactsSpy).toHaveBeenCalledTimes(2);
-    expect(removeObsoleteHookArtifactsSpy).toHaveBeenCalledWith(
-      '/proj/alpha',
-      migration.OBSOLETE_A3S_MARKER,
+    expect(removeObsoleteHookArtifactsSpy).toHaveBeenCalledWith('/proj/alpha');
+    expect(removeObsoleteHookArtifactsSpy).toHaveBeenCalledWith('/proj/beta');
+  });
+});
+
+describe('migrateHookScripts', () => {
+  let testDir: string;
+
+  beforeEach(() => {
+    setMockUi(true);
+    testDir = join(tmpdir(), `sonar-cli-migration-test-${Date.now()}`);
+    fs.mkdirSync(testDir, { recursive: true });
+  });
+
+  afterEach(() => {
+    setMockUi(false);
+    fs.rmSync(testDir, { recursive: true, force: true });
+  });
+
+  function writeOldScript(filename: string, content: string): string {
+    const dir = join(testDir, '.claude', 'hooks', 'sonar-secrets', 'build-scripts');
+    fs.mkdirSync(dir, { recursive: true });
+    const path = join(dir, filename);
+    fs.writeFileSync(path, content, 'utf-8');
+    return path;
+  }
+
+  it('rewrites sonar analyze to sonar analyze secrets in Unix scripts', () => {
+    const scriptPath = writeOldScript(
+      'pretool-secrets.sh',
+      '#!/bin/bash\nsonar analyze --file "$filePath" > /dev/null 2>&1\n',
     );
-    expect(removeObsoleteHookArtifactsSpy).toHaveBeenCalledWith(
-      '/proj/beta',
-      migration.OBSOLETE_A3S_MARKER,
+
+    migration.migrateHookScripts(testDir);
+
+    const content = fs.readFileSync(scriptPath, 'utf-8');
+    expect(content).toContain('sonar analyze secrets');
+    expect(content).not.toContain('sonar analyze --file');
+  });
+
+  it('rewrites all four hook script variants', () => {
+    const scripts = {
+      'pretool-secrets.sh': '#!/bin/bash\nsonar analyze --file "$f"\n',
+      'prompt-secrets.sh': '#!/bin/bash\nsonar analyze --file "$f"\n',
+      'pretool-secrets.ps1': '& sonar analyze --file $f\n',
+      'prompt-secrets.ps1': '& sonar analyze --file $f\n',
+    };
+
+    const paths: Record<string, string> = {};
+    for (const [name, content] of Object.entries(scripts)) {
+      paths[name] = writeOldScript(name, content);
+    }
+
+    migration.migrateHookScripts(testDir);
+
+    for (const [, path] of Object.entries(paths)) {
+      const content = fs.readFileSync(path, 'utf-8');
+      expect(content).toContain('sonar analyze secrets');
+      expect(content).not.toContain('sonar analyze --file');
+    }
+  });
+
+  it('uses globalDir over projectRoot when provided', () => {
+    const globalDir = join(testDir, 'global');
+    const dir = join(globalDir, '.claude', 'hooks', 'sonar-secrets', 'build-scripts');
+    fs.mkdirSync(dir, { recursive: true });
+    const scriptPath = join(dir, 'pretool-secrets.sh');
+    fs.writeFileSync(scriptPath, '#!/bin/bash\nsonar analyze --file "$f"\n', 'utf-8');
+
+    migration.migrateHookScripts(testDir, globalDir);
+
+    const content = fs.readFileSync(scriptPath, 'utf-8');
+    expect(content).toContain('sonar analyze secrets');
+    expect(content).not.toContain('sonar analyze --file');
+  });
+
+  it('leaves already-migrated scripts unchanged', () => {
+    const migrated = '#!/bin/bash\nsonar analyze secrets "$file_path"\n';
+    const scriptPath = writeOldScript('pretool-secrets.sh', migrated);
+
+    migration.migrateHookScripts(testDir);
+
+    const content = fs.readFileSync(scriptPath, 'utf-8');
+    expect(content).toBe(migrated);
+  });
+
+  it('completes without error when hook scripts do not exist on disk', () => {
+    // testDir has no hook scripts
+    expect(() => migration.migrateHookScripts(testDir)).not.toThrow();
+  });
+
+  it('logs debug and continues when a hook script cannot be read (read error)', () => {
+    // Create a directory where the script path is expected to be a file
+    // (causes readFileSync to throw EISDIR), exercising the catch branch
+    const secretsDir = join(testDir, '.claude', 'hooks', 'sonar-secrets', 'build-scripts');
+    fs.mkdirSync(secretsDir, { recursive: true });
+    fs.mkdirSync(join(secretsDir, 'pretool-secrets.sh'), { recursive: true });
+
+    expect(() => migration.migrateHookScripts(testDir)).not.toThrow();
+  });
+});
+
+describe('removeObsoleteHookArtifacts', () => {
+  let testDir: string;
+
+  beforeEach(() => {
+    setMockUi(true);
+    testDir = join(tmpdir(), `sonar-cli-migration-test-${Date.now()}`);
+    fs.mkdirSync(testDir, { recursive: true });
+  });
+
+  afterEach(() => {
+    setMockUi(false);
+    fs.rmSync(testDir, { recursive: true, force: true });
+  });
+
+  it('deletes the obsolete sonar-a3s hook directory', async () => {
+    const a3sDir = join(testDir, '.claude', 'hooks', OBSOLETE_A3S_MARKER, 'build-scripts');
+    fs.mkdirSync(a3sDir, { recursive: true });
+
+    await migration.removeObsoleteHookArtifacts(testDir);
+
+    expect(fs.existsSync(join(testDir, '.claude', 'hooks', OBSOLETE_A3S_MARKER))).toBe(false);
+  });
+
+  it('removes settings.json entries whose command references the marker and keeps others', async () => {
+    const claudeDir = join(testDir, '.claude');
+    fs.mkdirSync(claudeDir, { recursive: true });
+    const settingsPath = join(claudeDir, 'settings.json');
+    fs.writeFileSync(
+      settingsPath,
+      JSON.stringify(
+        {
+          hooks: {
+            PostToolUse: [
+              { hooks: [{ command: `sonar hook ${OBSOLETE_A3S_MARKER} run` }] },
+              { hooks: [{ command: 'sonar hook sonar-secrets run' }] },
+            ],
+          },
+        },
+        null,
+        2,
+      ),
+      'utf-8',
     );
+
+    await migration.removeObsoleteHookArtifacts(testDir);
+
+    const settings = JSON.parse(fs.readFileSync(settingsPath, 'utf-8')) as {
+      hooks: { PostToolUse: { hooks: { command: string }[] }[] };
+    };
+    const commands = settings.hooks.PostToolUse.flatMap((e) => e.hooks.map((h) => h.command));
+    expect(commands).toEqual(['sonar hook sonar-secrets run']);
+  });
+
+  it('does not throw when the settings file does not exist', async () => {
+    const result = await migration.removeObsoleteHookArtifacts(testDir);
+    expect(result).toBeUndefined();
+  });
+});
+
+describe('cleanObsoleteFromState', () => {
+  it('removes sonar-a3s from legacy hooks.installed', () => {
+    const state = getDefaultState('test');
+    state.agents['claude-code'].hooks.installed.push({
+      name: OBSOLETE_A3S_MARKER,
+      type: 'PostToolUse',
+      installedAt: new Date().toISOString(),
+    });
+
+    migration.cleanObsoleteFromState(state);
+
+    expect(
+      state.agents['claude-code'].hooks.installed.some((h) => h.name === OBSOLETE_A3S_MARKER),
+    ).toBe(false);
+  });
+
+  it('removes sonar-a3s from agentExtensions', () => {
+    const state = getDefaultState('test');
+    seedAgentExtension(state, {
+      id: 'a3s-ext',
+      agentId: 'claude-code',
+      projectRoot: '/some/project',
+      global: false,
+      kind: 'hook',
+      name: OBSOLETE_A3S_MARKER,
+      hookType: 'PostToolUse',
+      updatedByCliVersion: OLD_VERSION,
+      updatedAt: new Date().toISOString(),
+    });
+
+    migration.cleanObsoleteFromState(state);
+
+    expect(state.agentExtensions.some((e) => e.name === OBSOLETE_A3S_MARKER)).toBe(false);
+  });
+
+  it('does not remove unrelated entries from legacy hooks.installed', () => {
+    const state = getDefaultState('test');
+    state.agents['claude-code'].hooks.installed.push(
+      {
+        name: OBSOLETE_A3S_MARKER,
+        type: 'PostToolUse',
+        installedAt: new Date().toISOString(),
+      },
+      { name: 'sonar-secrets', type: 'PreToolUse', installedAt: new Date().toISOString() },
+    );
+
+    migration.cleanObsoleteFromState(state);
+
+    expect(
+      state.agents['claude-code'].hooks.installed.some((h) => h.name === 'sonar-secrets'),
+    ).toBe(true);
+  });
+
+  it('does not remove unrelated entries from agentExtensions', () => {
+    const state = getDefaultState('test');
+    seedAgentExtension(state, {
+      id: 'a3s-ext',
+      agentId: 'claude-code',
+      projectRoot: '/some/project',
+      global: false,
+      kind: 'hook',
+      name: OBSOLETE_A3S_MARKER,
+      hookType: 'PostToolUse',
+      updatedByCliVersion: OLD_VERSION,
+      updatedAt: new Date().toISOString(),
+    });
+    seedAgentExtension(state, {
+      id: 'secrets-ext',
+      agentId: 'claude-code',
+      projectRoot: '/some/project',
+      global: false,
+      kind: 'hook',
+      name: 'sonar-secrets',
+      hookType: 'PreToolUse',
+      updatedByCliVersion: OLD_VERSION,
+      updatedAt: new Date().toISOString(),
+    });
+
+    migration.cleanObsoleteFromState(state);
+
+    const survivors = state.agentExtensions.filter(
+      (e): e is HookExtension => e.kind === 'hook' && e.name === 'sonar-secrets',
+    );
+    expect(survivors).toHaveLength(1);
   });
 });
